refactor: Consolidate 5 error-handling files into errors/ package (issue #1097)
Created backend/src/errors/ package with:
- classifier.py: ErrorType, ErrorSeverity, ErrorClassification (from error_classifier.py)
- models.py: RecoveryStrategy, RecoveryAction, ErrorPattern (from error_patterns.py)
- handler.py: Merged error_handler.py + error_handlers.py (retry logic + FastAPI handlers)
- recovery.py: ErrorSupervisor, RecoveryResult (from error_recovery.py)
- __init__.py: Re-exports all public symbols for backward compatibility

Updated 4 legacy files in services/ to re-export from new errors/ package.
No breaking changes - all imports remain functional.
