+++ backend/src/services/error_handlers.py
@@ -1,570 +1,61 @@
 """
-Custom Exception Handlers for ModPorter AI Backend
+Error Handlers - backward compatibility re-export.
 
-Provides comprehensive error handling with user-friendly messages,
-graceful degradation, and structured logging.
-
-Issue: #455 - Comprehensive Error Handling (Phase 3)
+This module now imports from errors.handler.
 """
 
-import logging
-import os
-import traceback
-from typing import Any, Dict, Optional
-from fastapi import HTTPException, Request, status
-from fastapi.responses import JSONResponse
-from fastapi.exceptions import RequestValidationError
-from pydantic import BaseModel, ValidationError
-from datetime import datetime, timezone
-import uuid
-
-from .structured_logging import correlation_id_var, set_correlation_id
-
-
-# Check if running in debug mode (use function for dynamic checking)
-def is_debug_mode() -> bool:
-    """Check if debug mode is enabled."""
-    return os.getenv("DEBUG", "false").lower() == "true"
-
-
-# For backwards compatibility, also provide the constant (deprecated)
-DEBUG_MODE = is_debug_mode()
-
-# Configure logging
-logger = logging.getLogger(__name__)
-
-# Import metrics for error tracking (Issue #455)
-try:
-    from .metrics import record_error
-
-    METRICS_AVAILABLE = True
-except ImportError:
-    METRICS_AVAILABLE = False
-
-
-class ErrorResponse(BaseModel):
-    """Structured error response model"""
-
-    error_id: str
-    error_code: str  # Short uppercase error code: INVALID_FILE, PARSE_ERROR, etc.
-    error_type: str
-    error_category: str  # parse_error, asset_error, logic_error, etc.
-    message: str
-    user_message: str
-    is_retryable: bool  # Whether the client can retry this operation
-    timestamp: str
-    path: Optional[str] = None
-    method: Optional[str] = None
-    details: Optional[Dict[str, Any]] = None
-    correlation_id: Optional[str] = None  # correlation ID for tracing
-
-
-class ModPorterException(Exception):
-    """Base exception for ModPorter-specific errors"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "An error occurred. Please try again.",
-        error_type: str = "internal_error",
-        status_code: int = status.HTTP_500_INTERNAL_SERVER_ERROR,
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        self.message = message
-        self.user_message = user_message
-        self.error_type = error_type
-        self.status_code = status_code
-        self.details = details or {}
-        super().__init__(self.message)
-
-
-class ConversionException(ModPorterException):
-    """Exception raised during mod conversion"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Conversion failed. Please check your mod file and try again.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="conversion_error",
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            details=details,
-        )
-
-
-class FileProcessingException(ModPorterException):
-    """Exception raised during file processing"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Failed to process file. Please ensure it's a valid mod file.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="file_processing_error",
-            status_code=status.HTTP_400_BAD_REQUEST,
-            details=details,
-        )
-
-
-class ValidationException(ModPorterException):
-    """Exception raised during validation"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Validation failed. Please check your input.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="validation_error",
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            details=details,
-        )
-
-
-class NotFoundException(ModPorterException):
-    """Exception raised when a resource is not found"""
-
-    def __init__(self, resource: str, resource_id: str, user_message: Optional[str] = None):
-        super().__init__(
-            message=f"{resource} not found: {resource_id}",
-            user_message=user_message or f"Requested {resource} not found.",
-            error_type="not_found_error",
-            status_code=status.HTTP_404_NOT_FOUND,
-            details={"resource": resource, "resource_id": resource_id},
-        )
-
-
-# Error categories matching Issue #455
-ERROR_CATEGORIES = {
-    "parse_error": "Failed to parse the mod file. Please check the file format.",
-    "asset_error": "Error processing mod assets. Some assets may be missing.",
-    "logic_error": "An internal conversion error occurred. Please try again.",
-    "package_error": "Error packaging the converted mod. Please try again.",
-    "validation_error": "Validation failed. Please check your input.",
-    "network_error": "Network error. Please check your connection and try again.",
-    "rate_limit_error": "Too many requests. Please wait and try again.",
-    "timeout_error": "Operation timed out. Please try again.",
-    "unknown_error": "An unexpected error occurred. Please try again later.",
-    # Additional error types
-    "conversion_error": "Conversion failed. Please check your mod file.",
-    "http_error": "An error occurred while processing your request.",
-}
-
-# Error codes - short uppercase identifiers for client handling
-ERROR_CODES = {
-    "parse_error": "PARSE_ERROR",
-    "asset_error": "ASSET_ERROR",
-    "logic_error": "LOGIC_ERROR",
-    "package_error": "PACKAGE_ERROR",
-    "validation_error": "VALIDATION_ERROR",
-    "network_error": "NETWORK_ERROR",
-    "rate_limit_error": "RATE_LIMIT_ERROR",
-    "timeout_error": "TIMEOUT_ERROR",
-    "unknown_error": "UNKNOWN_ERROR",
-    "conversion_error": "CONVERSION_ERROR",
-    "http_error": "HTTP_ERROR",
-    "file_processing_error": "FILE_PROCESSING_ERROR",
-    "internal_error": "INTERNAL_ERROR",
-}
-
-# Retryable error categories - can be safely retried by client
-RETRYABLE_ERROR_CATEGORIES = {
-    "parse_error",
-    "asset_error",
-    "package_error",
-    "network_error",
-    "rate_limit_error",
-    "timeout_error",
-}
-
-# Non-retryable error categories - should not be retried
-NON_RETRYABLE_ERROR_CATEGORIES = {
-    "logic_error",
-    "validation_error",
-    "unknown_error",
-    "conversion_error",
-    "http_error",
-    "file_processing_error",
-    "internal_error",
-}
-
-
-class RateLimitException(ModPorterException):
-    """Exception raised when rate limit is exceeded"""
-
-    def __init__(
-        self,
-        message: str = "Rate limit exceeded",
-        user_message: str = "Too many requests. Please wait a moment and try again.",
-        retry_after: Optional[int] = None,
-    ):
-        details = {"retry_after": retry_after} if retry_after else {}
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="rate_limit_error",
-            status_code=status.HTTP_429_TOO_MANY_REQUESTS,
-            details=details,
-        )
-
-
-# New error categories for Issue #455
-class ParseError(ModPorterException):
-    """Error during file parsing"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Failed to parse the mod file. Please check the file format.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="parse_error",
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            details=details,
-        )
-
-
-class AssetError(ModPorterException):
-    """Error with asset processing"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Error processing mod assets. Some assets may be missing.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="asset_error",
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            details=details,
-        )
-
-
-class LogicError(ModPorterException):
-    """Error in conversion logic"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "An internal conversion error occurred. Please try again.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="logic_error",
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            details=details,
-        )
-
-
-class PackageError(ModPorterException):
-    """Error during packaging"""
-
-    def __init__(
-        self,
-        message: str,
-        user_message: str = "Error packaging the converted mod. Please try again.",
-        details: Optional[Dict[str, Any]] = None,
-    ):
-        super().__init__(
-            message=message,
-            user_message=user_message,
-            error_type="package_error",
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            details=details,
-        )
-
-
-def _categorize_error(error: Exception) -> str:
-    """
-    Categorize an error based on its type.
-
-    Returns one of: parse_error, asset_error, logic_error,
-                    package_error, validation_error, network_error,
-                    rate_limit_error, timeout_error, unknown_error
-    """
-    error_type = type(error).__name__
-    error_msg = str(error).lower()
-
-    # Check specific error types
-    if isinstance(error, ParseError):
-        return "parse_error"
-    if isinstance(error, AssetError):
-        return "asset_error"
-    if isinstance(error, LogicError):
-        return "logic_error"
-    if isinstance(error, PackageError):
-        return "package_error"
-    if isinstance(error, ValidationError):
-        return "validation_error"
-    if isinstance(error, RateLimitException):
-        return "rate_limit_error"
-    # Import ConversionError here to avoid circular imports
-    from .error_handler import ConversionError
-
-    if isinstance(error, ConversionError):
-        return "conversion_error"
-
-    # Check error message patterns
-    if "parse" in error_type.lower() or "parse" in error_msg:
-        return "parse_error"
-    if "asset" in error_type.lower() or "asset" in error_msg:
-        return "asset_error"
-    if "logic" in error_type.lower() or "convert" in error_msg:
-        return "logic_error"
-    if "package" in error_type.lower() or "packag" in error_msg:
-        return "package_error"
-    if "valid" in error_type.lower() or "validation" in error_msg or "invalid" in error_msg:
-        return "validation_error"
-    if "network" in error_type.lower() or "connection" in error_msg:
-        return "network_error"
-    if "rate limit" in error_msg or "429" in error_msg:
-        return "rate_limit_error"
-    if "timeout" in error_type.lower() or "timeout" in error_msg:
-        return "timeout_error"
-
-    return "unknown_error"
-
-
-def create_error_response(
-    error: Exception, request: Request, include_traceback: bool = False
-) -> ErrorResponse:
-    """Create a structured error response
-
-    NOTE: Sensitive error details are NEVER included in HTTP responses.
-    - Stack traces are logged server-side only, never sent to clients
-    - Exception messages are sanitized to generic user messages
-    - Internal details are sanitized unless DEBUG_MODE is enabled
-    """
-    error_id = str(uuid.uuid4())[:8]
-    timestamp = datetime.now(timezone.utc).isoformat()
-
-    # Get or create correlation ID
-    correlation_id = correlation_id_var.get()
-    if not correlation_id:
-        correlation_id = set_correlation_id()
-
-    # Determine error type, category and user message
-    if isinstance(error, ModPorterException):
-        error_type = error.error_type
-        error_category = _categorize_error(error)
-        user_message = error.user_message
-        message = error.message
-        # Only include error details in debug mode
-        details = error.details if is_debug_mode() else {}
-    elif isinstance(error, HTTPException):
-        error_type = "http_error"
-        error_category = _categorize_error(error)
-        user_message = error.detail
-        message = str(error.detail)
-        # Only include status code in production
-        details = {"status_code": error.status_code} if is_debug_mode() else {}
-    elif isinstance(error, RequestValidationError):
-        error_type = "validation_error"
-        error_category = "validation_error"
-        user_message = "Invalid request data. Please check your input."
-        message = str(error)
-        # In production, don't expose validation error details
-        details = {"errors": error.errors()} if is_debug_mode() else {}
-    else:
-        error_type = "internal_error"
-        error_category = _categorize_error(error)
-        user_message = ERROR_CATEGORIES.get(
-            error_category, "An unexpected error occurred. Please try again later."
-        )
-        # Only expose message in debug mode - never expose raw exception to clients
-        message = str(error) if is_debug_mode() else "An internal server error occurred"
-        details = {}
-
-    # Add traceback only in debug mode (NEVER in production)
-    if include_traceback and is_debug_mode():
-        details["traceback"] = traceback.format_exc()
-
-    # Determine error code and retryability from category
-    error_code = ERROR_CODES.get(error_category, "UNKNOWN_ERROR")
-    is_retryable = error_category in RETRYABLE_ERROR_CATEGORIES
-
-    return ErrorResponse(
-        error_id=error_id,
-        error_code=error_code,
-        error_type=error_type,
-        error_category=error_category,
-        message=message,
-        user_message=user_message,
-        is_retryable=is_retryable,
-        timestamp=timestamp,
-        path=str(request.url.path),
-        method=request.method,
-        details=details,
-        correlation_id=correlation_id,
-    )
-
-
-def _record_error_metric(error_category: str, error_type: str, source: str = "api"):
-    """Record error metrics if metrics are available."""
-    if METRICS_AVAILABLE:
-        try:
-            record_error(error_category, error_type, source)
-        except Exception:
-            pass  # Don't let metrics errors affect normal flow
-
-
-async def modporter_exception_handler(request: Request, exc: ModPorterException) -> JSONResponse:
-    """Handler for ModPorter-specific exceptions"""
-    error_response = create_error_response(exc, request)
-    logger.error(
-        f"[{error_response.error_id}] {error_response.error_type}: {error_response.message}"
-    )
-
-    # Record error metric
-    _record_error_metric(
-        error_category=error_response.error_category,
-        error_type=error_response.error_type,
-        source="api",
-    )
-
-    return JSONResponse(status_code=exc.status_code, content=error_response.model_dump())
-
-
-async def http_exception_handler(request: Request, exc: HTTPException) -> JSONResponse:
-    """Handler for HTTP exceptions"""
-    error_response = create_error_response(exc, request)
-    logger.warning(f"[{error_response.error_id}] HTTP {exc.status_code}: {error_response.message}")
-
-    # Only record 4xx/5xx errors as metrics
-    if exc.status_code >= 400:
-        _record_error_metric(
-            error_category=error_response.error_category, error_type="HTTPException", source="api"
-        )
-
-    return JSONResponse(status_code=exc.status_code, content=error_response.model_dump())
-
-
-async def validation_exception_handler(
-    request: Request, exc: RequestValidationError
-) -> JSONResponse:
-    """Handler for validation exceptions"""
-    error_response = create_error_response(exc, request)
-    logger.warning(f"[{error_response.error_id}] Validation error: {error_response.message}")
-
-    # Record validation error metric
-    _record_error_metric(
-        error_category="validation_error", error_type="RequestValidationError", source="api"
-    )
-
-    return JSONResponse(
-        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, content=error_response.model_dump()
-    )
-
-
-async def generic_exception_handler(request: Request, exc: Exception) -> JSONResponse:
-    """Handler for all other exceptions
-
-    Full error details are logged server-side with traceback for debugging.
-    Clients only receive a generic error message without sensitive information.
-    """
-    # Always log the full exception with traceback (server-side only)
-    logger.error(f"[{exc.__class__.__name__}] Unhandled exception: {str(exc)}", exc_info=True)
-
-    # Create sanitized error response (no traceback exposed to clients)
-    error_response = create_error_response(exc, request, include_traceback=False)
-
-    # Record unhandled exception metric
-    _record_error_metric(
-        error_category=error_response.error_category, error_type=type(exc).__name__, source="api"
-    )
-
-    return JSONResponse(
-        status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content=error_response.model_dump()
-    )
-
-
-def register_exception_handlers(app):
-    """Register all exception handlers with the FastAPI app
-
-    Validates that all handlers are callable with the correct signature before registration.
-    This prevents issues where incorrectly defined handlers cause runtime errors.
-    """
-    import inspect
-    from fastapi.exceptions import RequestValidationError
-
-    # Define handlers to register with their expected signatures
-    handlers_to_register = [
-        (ModPorterException, modporter_exception_handler, "ModPorterException"),
-        (HTTPException, http_exception_handler, "HTTPException"),
-        (RequestValidationError, validation_exception_handler, "RequestValidationError"),
-        (Exception, generic_exception_handler, "Exception"),
-    ]
-
-    for exception_class, handler_func, handler_name in handlers_to_register:
-        # Validate handler is callable
-        if not callable(handler_func):
-            raise TypeError(
-                f"Cannot register handler '{handler_name}': not callable. "
-                f"Got type: {type(handler_func)}"
-            )
-
-        # Validate handler has the correct signature (request, exc)
-        try:
-            sig = inspect.signature(handler_func)
-            params = list(sig.parameters.keys())
-            if len(params) < 2:
-                raise TypeError(
-                    f"Cannot register handler '{handler_name}': expected at least 2 parameters "
-                    f"(request, exc), but handler has {len(params)} parameter(s): {params}. "
-                    f"Full signature: {sig}"
-                )
-        except ValueError as e:
-            # Some callable objects don't support inspect.signature
-            # In this case, we'll try to register anyway and let FastAPI handle it
-            pass
-
-        # Register the handler
-        app.add_exception_handler(exception_class, handler_func)
-
-
-def verify_exception_handlers(app) -> Dict[str, bool]:
-    """Verify that exception handlers are properly registered.
-
-    Returns a dictionary showing the registration status of each handler type.
-
-    Args:
-        app: The FastAPI application instance to verify.
-
-    Returns:
-        Dictionary mapping handler names to their registration status.
-    """
-    from fastapi.exceptions import RequestValidationError
-
-    handlers_to_check = [
-        ("ModPorterException", ModPorterException),
-        ("HTTPException", HTTPException),
-        ("RequestValidationError", RequestValidationError),
-        ("Generic Exception", Exception),
-    ]
-
-    result = {}
-    for name, exc_class in handlers_to_check:
-        # Check if there's a handler for this exception class
-        # FastAPI stores handlers in exception_handlers dict
-        has_handler = exc_class in getattr(app, "exception_handlers", {})
-        result[name] = has_handler
-
-    return result
+from errors.handler import (
+    ErrorResponse,
+    ModPorterException,
+    ConversionException,
+    FileProcessingException,
+    ValidationException,
+    NotFoundException,
+    RateLimitException,
+    ParseError,
+    AssetError,
+    LogicError,
+    PackageError,
+    ERROR_CATEGORIES,
+    ERROR_CODES,
+    RETRYABLE_ERROR_CATEGORIES,
+    NON_RETRYABLE_ERROR_CATEGORIES,
+    _categorize_error,
+    create_error_response,
+    register_exception_handlers,
+    verify_exception_handlers,
+    modporter_exception_handler,
+    http_exception_handler,
+    validation_exception_handler,
+    generic_exception_handler,
+    is_debug_mode,
+    DEBUG_MODE,
+)
+
+__all__ = [
+    "ErrorResponse",
+    "ModPorterException",
+    "ConversionException",
+    "FileProcessingException",
+    "ValidationException",
+    "NotFoundException",
+    "RateLimitException",
+    "ParseError",
+    "AssetError",
+    "LogicError",
+    "PackageError",
+    "ERROR_CATEGORIES",
+    "ERROR_CODES",
+    "RETRYABLE_ERROR_CATEGORIES",
+    "NON_RETRYABLE_ERROR_CATEGORIES",
+    "_categorize_error",
+    "create_error_response",
+    "register_exception_handlers",
+    "verify_exception_handlers",
+    "modporter_exception_handler",
+    "http_exception_handler",
+    "validation_exception_handler",
+    "generic_exception_handler",
+    "is_debug_mode",
+    "DEBUG_MODE",
+]
